sync with Parsoid (6700f2673)
Convert `extTagWidths` to `extTagOffsets`. This allows us to uniformly 
update the offsets between character encodings. Previously it was 
impossible to fix up extTagWidths[1] on the TagTk without locating the 
tsr for the matching EndTagTk; in general we can't update a width 
without knowing what part of the original string it corresponds to, so 
we can tell how many bytes each of the original
characters uses.